// ng/src/app/shared/interfaces/user-detail.ts
export interface UserDetail {
	userDetailId: string;
	userDetailUserId: string;
	userDetailAboutMe: string;
	userDetailAge: string;
	userDetailCareer: string;
	userDetailDisplayEmail: string;
	userDetailEducation: string;
	userDetailGender: string;
	userDetailInterests: string;
	userDetailRace: string;
	userDetailReligion: string;
}